moved the repl as its not really part of the core language
